Customizing models, collect all settings in blog.settings, simple documentation notes
